+++ src/app.module.ts
@@ -23,6 +23,7 @@ import { NotificationModule } from './modules/notification/notification.module';
 import { JobModule } from './modules/job/job.module';
 import { MaintenanceModule } from './modules/maintenance/maintenance.module';
 import { ManagedDatabaseModule } from './modules/managed-database/managed-database.module';
+import { PanelDomainModule } from './modules/panel-domain/panel-domain.module';
 import { ProjectModule } from './modules/project/project.module';
 import { ProjectTransferModule } from './modules/project-transfer/project-transfer.module';
 import { ProxyModule } from './modules/proxy/proxy.module';
@@ -63,6 +64,7 @@ import { VolumeBackupModule } from './modules/volume-backup/volume-backup.module
     RegistryModule,
     ApplicationModule,
     ProxyModule,
+    PanelDomainModule,
     GitCredentialModule,
     ManagedDatabaseModule,
     BackupDestinationModule,